requirements.txt
flask 
flask-bcrypt 
flask-cors 
torch 
transformers 
google-generativeai
whisper 
scipy 
sounddevice 
camelot-py[cv]
PyPDF2 
pdf2image 
pdfplumber 
pytesseract 
pillow 
accelerate
transformers
qdrant-client 
firebase-admin 
python-dotenv 
requests 
numpy

